jk_img_classification/requirements.txt
attrs==20.3.0
awscli==1.18.223
botocore==1.19.63
certifi==2020.12.5
chainer==7.7.0
chardet==4.0.0
colorama==0.4.3
cycler==0.10.0
docutils==0.15.2
facenet-pytorch @ file:///home/sumant/Desktop/temp/facenet-pytorch
filelock==3.0.12
future==0.18.2
idna==2.10
importlib-metadata==3.4.0
iniconfig==1.1.1
jmespath==0.10.0
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.18.4
onnx==1.8.1
onnxruntime==1.6.0
packaging==20.9
pandas==1.1.5
Pillow==7.1.2
pluggy==0.13.1
protobuf==3.14.0
pt-image-classify @ file:///media/sumant/projects/python/company_wise/safehalo/safe_swipe_v2_artifacts/pytorch_image_classification_artifacts/pytorch_image_classification
pt-img-classify @ file:///media/sumant/projects/python/company_wise/safehalo/safe_swipe_v2_artifacts/pytorch_image_classification_artifacts/pytorch_image_classification
py==1.10.0
pyasn1==0.4.8
pyparsing==2.4.7
pytest==6.2.2
python-dateutil==2.8.1
pytz==2020.5
PyYAML==5.3.1
requests==2.25.1
rsa==4.5
s3transfer==0.3.4
six==1.15.0
toml==0.10.2
torch==1.5.0+cpu
torch-summary==1.4.5
torchvision==0.6.0+cpu
tqdm==4.56.0
typing-extensions==3.7.4.3
urllib3==1.26.2
zipp==3.4.0
